refactor route listener
